(RAZOR-1065) Add `configdir` setting
This commit adds a `configdir` setting to the razor-server and pe-razor-server
projects to determine where to install the `config.yaml` file. For PE, this
file is managed by the puppetlabs-pe_razor module, so we place it in the razor
install_root. Otherwise, we place it in the sysconfdir (under /etc), where we
expect user-managed files to live. Since the shiro.ini file is always managed
by the user, it always lives under /etc, so we use the previously-defined
`sysconfdir` for that file.