(RAZOR-1065) Add conditional Java requirements
This commit adds a boolean `pe_package` setting to the razor-server and
pe-razor-server projects and uses this setting to conditionally require pe-java
if we are building pe-razor-server. This commit also adds install commands for
adding Java-related environment variables to the service files so that pe-java
is used when running pe-razor-server.