(maint) Add (empty) deb_targets
This commit adds a deb_targets key to build_defaults.yaml. When promoting to PE,
the signing task splits on deb (and rpm) targets, so even though we don't ship
pe-razor-server for debian, we need to include some value for the key so that
we don't try to split on nil.