(RAZOR-1082) Don't use /tmp in scriptlets
This commit changes the directory we use for moving config files out of the way
during upgrade. Previously, we created a directory under /tmp, but because /tmp
is world writeable, this was a security vulnerability. We now simply rename the
config file pre-upgrade and rename it back afterwards, maintaining the same
permissions that the config directories have in the first place.