(SEC-166) Dust off razor packaging to get a security update out
This commit updates the razor vanagon project to look for packages at artifactory. It also updates razor to install pe-java instead of pe-razor-java which was a band-aid to prevent java 11 incompatability which is not an issue in irving. NOTE: given we only need to ship this for irving and there is only a single streem for razor this is acceptable. This is only safe because we are no longer shipping razor with modern PE.